+++ doctor-event-api/src/main/java/io/terminus/doctor/event/model/DoctorReportBoar.java
@@ -84,7 +84,7 @@ public class DoctorReportBoar implements Serializable {
     /**
      * 日均存栏
      */
-    private Integer dailyPigCount;
+    private Double dailyPigCount;
     
     /**
      * 期末存栏
